tests/requirements-test.txt
# Agent Zero Test Dependencies
# Install with: pip install -r tests/requirements-test.txt

# Core testing framework
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-xdist>=3.5.0        # Parallel test execution
pytest-timeout>=2.2.0      # Test timeouts
pytest-mock>=3.12.0        # Enhanced mocking

# HTTP mocking and recording
vcrpy>=6.0.0               # Record/replay HTTP calls
pytest-recording>=0.13.0   # VCR.py pytest integration
responses>=0.25.0          # HTTP mocking

# LLM evaluation
deepeval>=0.21.0           # LLM evaluation metrics

# Snapshot testing
syrupy>=4.6.0              # Snapshot assertions

# Property-based testing
hypothesis>=6.100.0        # Property-based testing

# Utilities
dirty-equals>=0.7.0        # Flexible assertions
freezegun>=1.4.0           # Time mocking
faker>=24.0.0              # Test data generation

# Performance testing (optional)
pytest-benchmark>=4.0.0    # Benchmarking
